remove duplicate code step 5 - move up utils independent from privates
